Some updates

+++ web/pkg/provision.go
@@ -5,17 +5,15 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
-	"strings"
 	"time"
 )
-
-const AssumeRolePostfix = "-assume-role-id"
 
 // const KeepRolesFor = time.Hour * 24
 const KeepRolesFor = time.Hour * 24
 
 type ProvisionRoleRequest struct {
-	RolePrefix string `json:"role_prefix"`
+	RoleName          string `json:"role_name"`
+	RequireExternalId bool
 }
 
 type ProvisionRoleOutput struct {
@@ -30,10 +28,9 @@ func ProvisionRole(ctx *Context, client *iam.Client, req *ProvisionRoleRequest) 
 		}
 	}()
 
-	role, err := client.CreateRole(ctx, &iam.CreateRoleInput{
-		RoleName:    aws.String(req.RolePrefix + RandStringRunes(24) + AssumeRolePostfix),
-		Description: aws.String("role for assume-role-id"),
-		AssumeRolePolicyDocument: aws.String(`{
+	var policy string
+	if req.RequireExternalId {
+		policy = `{
 			"Version": "2012-10-17",
 			"Statement": [	
 				{	
@@ -44,7 +41,30 @@ func ProvisionRole(ctx *Context, client *iam.Client, req *ProvisionRoleRequest) 
 					"Action": "sts:AssumeRole"	
 				}
 			]
-		}`),
+		}`
+	} else {
+		policy = `{
+			"Version": "2012-10-17",
+			"Statement": [	
+				{	
+					"Effect": "Allow",
+					"Principal": {
+						"AWS": "*"
+					},
+					"Action": "sts:AssumeRole",
+					"Condition": {
+						"Null": {
+							"sts:ExternalId": "false"
+						}
+					}
+				}
+			]
+		}`
+	}
+	role, err := client.CreateRole(ctx, &iam.CreateRoleInput{
+		RoleName:                 aws.String(req.RoleName),
+		Description:              aws.String("role for assume-role-id"),
+		AssumeRolePolicyDocument: aws.String(policy),
 		Tags: []types.Tag{
 			{
 				Key:   aws.String("assume-role-id"),
@@ -56,28 +76,45 @@ func ProvisionRole(ctx *Context, client *iam.Client, req *ProvisionRoleRequest) 
 		return nil, fmt.Errorf("creating role: %w", err)
 	}
 
+	// Shouldn't matter much but just to be safe.
+	if _, err := client.AttachRolePolicy(ctx, &iam.AttachRolePolicyInput{
+		RoleName:  role.Role.RoleName,
+		PolicyArn: aws.String("arn:aws:iam::aws:policy/AWSDenyAll"),
+	}); err != nil {
+		return nil, fmt.Errorf("attaching policy: %w", err)
+	}
+
 	return &ProvisionRoleOutput{
 		RoleArn: *role.Role.Arn,
 	}, nil
 }
 
 func CleanUpOldRoles(ctx *Context, client *iam.Client) error {
-	roles := []types.Role{}
 	resp, err := client.ListRoles(ctx, &iam.ListRolesInput{})
 	if err != nil {
 		return fmt.Errorf("listing roles: %w", err)
 	}
 
-	for _, role := range resp.Roles {
-		if strings.HasSuffix(*role.RoleName, AssumeRolePostfix) {
-			roles = append(roles, role)
-		}
-	}
-
 	cutoff := time.Now().UTC().Add(-KeepRolesFor)
-	for _, role := range roles {
-		if role.CreateDate.UTC().Before(cutoff) {
+	for _, role := range resp.Roles {
+		if IsOurRole(role) && role.CreateDate.UTC().Before(cutoff) {
 			ctx.Debug.Printf("deleting role %s", *role.RoleName)
+
+			resp, err := client.ListAttachedRolePolicies(ctx, &iam.ListAttachedRolePoliciesInput{
+				RoleName: role.RoleName,
+			})
+			if err != nil {
+				return fmt.Errorf("listing attached policies %s: %w", *role.RoleName, err)
+			}
+
+			for _, policy := range resp.AttachedPolicies {
+				if _, err := client.DetachRolePolicy(ctx, &iam.DetachRolePolicyInput{
+					RoleName:  role.RoleName,
+					PolicyArn: policy.PolicyArn,
+				}); err != nil {
+					return fmt.Errorf("detaching policy %s: %w", *policy.PolicyArn, err)
+				}
+			}
 
 			if _, err := client.DeleteRole(ctx, &iam.DeleteRoleInput{
 				RoleName: role.RoleName,
@@ -88,4 +125,13 @@ func CleanUpOldRoles(ctx *Context, client *iam.Client) error {
 	}
 
 	return nil
+}
+
+func IsOurRole(role types.Role) bool {
+	for _, tag := range role.Tags {
+		if *tag.Key == "assume-role-id" && *tag.Value == "true" {
+			return true
+		}
+	}
+	return false
 }
